Remove try/catch & update text
